Move named element and namespace config to UML package
It's UML related.
