+++ gaphor/RAAML/stpa/stpatoolbox.py
@@ -1,17 +1,15 @@
 """The definition for the STPA section of the RAAML toolbox."""
 
-from gaphor.diagram.diagramtoolbox import (
-    ToolDef,
-    ToolSection,
-    default_namespace,
-    named_element_config,
-    namespace_config,
-    new_item_factory,
-)
+from gaphor.diagram.diagramtoolbox import ToolDef, ToolSection, new_item_factory
 from gaphor.i18n import gettext
 from gaphor.RAAML import diagramitems, raaml
 from gaphor.SysML import diagramitems as sysml_items
 from gaphor.UML import diagramitems as uml_items
+from gaphor.UML.toolboxconfig import (
+    default_namespace,
+    named_element_config,
+    namespace_config,
+)
 
 
 def loss_config(new_item):
