+++ gaphor/UML/actions/actionstoolbox.py
@@ -5,14 +5,10 @@ from gaphas.item import SE
 
 from gaphor import UML
 from gaphor.core import gettext
-from gaphor.diagram.diagramtoolbox import (
-    ToolDef,
-    ToolSection,
-    namespace_config,
-    new_item_factory,
-)
+from gaphor.diagram.diagramtoolbox import ToolDef, ToolSection, new_item_factory
 from gaphor.UML import diagramitems
 from gaphor.UML.recipes import owner_package
+from gaphor.UML.toolboxconfig import namespace_config
 
 
 def activity_config(new_item, name=None):
