+++ gaphor/UML/classes/classestoolbox.py
@@ -4,13 +4,9 @@ from gaphas.item import SE
 
 from gaphor import UML
 from gaphor.core import gettext
-from gaphor.diagram.diagramtoolbox import (
-    ToolDef,
-    ToolSection,
-    namespace_config,
-    new_item_factory,
-)
+from gaphor.diagram.diagramtoolbox import ToolDef, ToolSection, new_item_factory
 from gaphor.UML import diagramitems
+from gaphor.UML.toolboxconfig import namespace_config
 
 
 def composite_association_config(assoc_item: diagramitems.AssociationItem) -> None:
